% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_test.R
\name{batch_test}
\alias{batch_test}
\title{a test thingy}
\usage{
batch_test(rep, file = "/work/pi_cschweik_umass_edu/batch_test/btest/test_")
}
\arguments{
\item{rep}{Rep number}

\item{file}{File to write to}
}
\description{
Rep 2 will throw an error, and rep 4 will blow out memory
}
