% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzy.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize user parameters when the package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{(required by R; not used)}

\item{pkgname}{(required by R; not used)}
}
\description{
Initialize user parameters when the package is loaded
}
\keyword{internal}
