% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_walk_path.R
\name{drive_walk_path}
\alias{drive_walk_path}
\title{Walk down a file path a folder at a time in the currently Google Drive}
\usage{
drive_walk_path(path)
}
\arguments{
\item{path}{drive path on Google Drive}
}
\value{
A dribble for full path. Returns NULL if the directory does not exist.
}
\description{
This is far, far faster than using drive_get or drive_ls with a deep path on a
drive with LOTS of files.
}
\details{
The following takes 6-7 min on our Google Drive, which has ~16,000 files

\code{drive_ls(path = 'UAS Data Collection/Westport/RFM Processing Inputs/Orthomosaics/')}

while this call takes 6 seconds

\code{drive_walk_path(path = 'UAS Data Collection/Westport/RFM Processing Inputs/Orthomosaics/')}
}
\keyword{internal}
