% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_jobs.R
\name{filter_jobs}
\alias{filter_jobs}
\title{Filter jobs database}
\usage{
filter_jobs(filter)
}
\arguments{
\item{filter}{A named list to filter jobs with. List items are \verb{<field in jdb> = <value>}, where \if{html}{\out{<value>}} is a regex for character fields, or an
actual value (or vector of values) for logical or numeric fields.}
}
\value{
A logical vector corresponding to rows in \code{the$jdb}.
}
\description{
Filter jobs database
}
