% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_standards.R
\name{find_standards}
\alias{find_standards}
\title{Find raster standard for each site that doesn't already have one. It will be the oldest Mica file for each site.}
\usage{
find_standards(
  subdirs = c("RFM processing inputs/Orthomosaics", "Share/Photogrammetry DEMs",
    "Share/Canopy height models"),
  basedir = "c:/Work/etc/saltmarsh/data",
  resultbase = "c:/Work/etc/saltmarsh/data"
)
}
\arguments{
\item{subdirs}{subdirectories to search, ending with slash. Default = orthos, DEMs, and canopy height models (okay to include empty or
nonexistent directories)}

\item{basedir}{full path to subdirs}

\item{resultbase}{base name of result base directory}
}
\description{
Requires \verb{<pars>sites.txt}, table of site abbreviation, site name, footprint shapefile, raster standard.
}
\details{
Result is written to \code{resultbase/sites.txt}, new version of sites file. ** Copy this to pars/sites.txt after running **.

Selected standards must be in EPSG:4326. A warning will be returned for files in other projections; you'll have to reproject
the source file or pick a standard manually for these sites.

I might abandon this and require picking standards by hand. There aren't many sites, so it's not a particular burden, and
human eyes might be helpful at this step.
}
\keyword{internal}
