% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit models}
\usage{
fit(
  site = the$site,
  datafile = the$datafile,
  method = "rf",
  vars = NULL,
  exclude = NULL,
  years = NULL,
  maxmissing = 0.05,
  top_importance = 20,
  reread = FALSE,
  holdout = 0.2,
  auc = TRUE
)
}
\arguments{
\item{site}{Site name (potentially a metasite for multi-site models); determines path
of data. Default = \code{the$site}; must be set here if \code{the$site} is NULL. \code{the$site}
will be set by the first call of \code{fit}; it may also be set directly.}

\item{datafile}{Name of data file. Extension \code{.RDS} must be included. Default =
\code{the$datafile}, or \code{data.RDS} if \code{the$datafile} is NULL.}

\item{method}{One of \code{rf} for Random Forest, \code{boost} for AdaBoost. Default = \code{rf}.}

\item{vars}{An optional vector of variables to restrict analysis to. Default = NULL,
all variables.}

\item{exclude}{An optional vector of variables to exclude.}

\item{years}{An optional vector of years to restrict variables to.}

\item{maxmissing}{Maximum proportion of missing training points allowed before a
variable is dropped.}

\item{top_importance}{Give number of variables to keep for variable importance.}

\item{reread}{If TRUE, forces reread of datafile.}

\item{holdout}{Proportion of points to hold out. For Random Forest, this specifies
the size of the single validation set, while for boosting, it is the size of each
of the testing and validation sets.}

\item{auc}{If TRUE, calculate class probabilities so we can calculate AUC.}
}
\description{
Fit models
}
