% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir.R
\name{get_dir}
\alias{get_dir}
\title{Return directory listing as a data frame from the local drive, Google Drive, or SFTP site}
\usage{
get_dir(path, sourcedrive = "local", logfile, sftp)
}
\arguments{
\item{path}{directory path (must end with '/' on Google Drive)}

\item{logfile}{log file, for reporting missing directories (which don't throw an error)}

\item{sftp}{list of url = address of site, user = credentials (optional)}
}
\description{
A sister function to get_file
}
\details{
Returns data frame with name (filenames), and id (Google Drive id, only if \verb{sourcedrive = }google`)
\itemize{
\item when \code{sourcedrive = local} name is full path to local files
\item when \code{sourcedrive = google}, name is just the base name
\item when \code{sourcedrive = sftp}, name is the full path to files on the SFTP site
\item When directories aren't found, the name is added to the log and NULL is returned
}

Notes:
\itemize{
\item paths for Google Drive are case-sensitive
\item initial runs with Google Drive in a session open the browser for authentication or wait
for input from the console, so don't run blindly when using the Google Drive
\item SFTP directory info is crazy. Apparently the format is highly platform-dependent. The targeted
use for this is running on Unity / fetching files from SFTP on the NAS that we don't have yet,
so I'll have to revisit this to tailor it to the specific platforms.
\item When testing this on my laptop / fetching files via SFTP from the Landeco cluster, I found two date issues:
\enumerate{
\item Seconds are truncated. I'm adding one minute so date checks won't fail. It seems safe to assume source
files won't be updated within a minute of downloading them.
\item Strangely, summer dates are reported differently by xplorer2 and DOS/Command Prompt (1:12 pm) vs. Win
Explorer, FileZilla, and this code (2:12 pm). Absolutely crazy. In the final setup, we'll have Linux
servers on both sides so hopefully this'll clear up.
}
}
}
