% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Return a file name from the local drive, Google Drive, or SFTP}
\usage{
get_file(name, gd, logfile)
}
\arguments{
\item{name}{File path and name}

\item{gd}{Source drive info, named list of
\itemize{
\item \code{dir} - Google directory info, from \link{get_dir}
\item \code{sourcedrive} - which source drive (\code{local}, \code{google}, or \code{sftp})
\item \code{sftp} - list(url, user)
\item \code{cachedir} - local cache directory
}}

\item{logfile}{Log file, for reporting missing directories (which don't throw an error)}
}
\value{
path to file on local drive
}
\description{
If reading from the Google Drive or SFTP, the file is cached on the scratch drive
(\code{gd$cache}), and reused as long as it isn't outdated. Get \code{gd$dir} with sister function \link{get_dir}.
}
\details{
Notes:
\itemize{
\item this code assumes that all files have unique names, even if from different directories. This
holds true for the UAS salt marsh project, so good enough. *** IF REPURPOSING THIS CODE, beware! ***
\item cached files are reused if they're not outdated. Downloads from Google Drive or SFTP to Unity are
wicked fast, so don't feel bad freeing up the scratch drive after a run. It's the polite thing
to do.
\item we don't check for a full cache drive, as the Unity scratch drive has a 50 TB limit and we
have < 1 TB of data. Again, IF REPURPOSING, beware!
\item we protect against crashed or interrupted downloads by downloading to a temporary file that
is renamed after completion.
\item when reading from SFTP, the entire file must be able to fit in memory. There should be plenty
of room for the files in the salt marsh project.
}
}
\keyword{internal}
