% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_id.R
\name{get_job_id}
\alias{get_job_id}
\title{Get the Slurm job id for the specified batchtools job id in the specified registry. If
the job.id doesn't exist, returns NA.}
\usage{
get_job_id(job.id, reg = getDefaultRegistry())
}
\arguments{
\item{job.id}{The job.id of a batchtools job}

\item{reg}{batchtools registry object}
}
\description{
In order for this to work, you must add the following line to your Slurm template
(slurm.tmpl): \cr \cr
\verb{echo $SLURM_JOB_ID > <\%= log.file \%>.slurmjobid}
}
