% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_status.R
\name{get_job_status}
\alias{get_job_status}
\title{Get the status for Slurm jobs}
\usage{
get_job_status(jobid = NULL, days = NULL)
}
\arguments{
\item{jobid}{Slurm JobID}

\item{days}{Number of days to look back}
}
\value{
Data frame with jobID, State, and Reason
}
\description{
Gets the Slurm State and Reason codes for a specific job or all jobs launched
in the past n days.
}
\details{
Either specify the \code{jobid} with the Slurm JobID to get the status for a
single job, or \code{days} to get the status for all jobs run in the past \code{days}
days. If you don't specify either, you'll get the Slurm default: jobs you
launched today.

There is overhead for the Slurm call, so if processing several jobs, best to
use an adequate value for days and pull info on your jobs from the result.

Only the primary \code{JobID}s are returned; \verb{_batch}, \verb{_extern}, and other
sidecar jobs are dropped.

\code{get_job_status} uses the Slurm \code{sacct} command, and results differ slightly
from those returned by \code{squeue}. Interacting with Slurm requires setting up
\code{ssh} to connect to a login node.
}
