% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources.R
\name{get_resources}
\alias{get_resources}
\title{Combine user-supplied resources with function-specific defaults}
\usage{
get_resources(user, func)
}
\arguments{
\item{user}{List of user-specified resources; NULL if none}

\item{func}{List of function-specified resources}
}
\value{
Combined list of resources, with user resources taking precedence
}
\description{
Combine user-supplied resources with function-specific defaults
}
