% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slurm_template.R
\name{get_slurm_template}
\alias{get_slurm_template}
\title{Get the path for the slurm template}
\usage{
get_slurm_template()
}
\value{
The path to the slurm template
}
\description{
Get the path for the slurm template
}
