% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Give info on batch jobs}
\usage{
info(what = "summary", sweep = TRUE)
}
\arguments{
\item{what}{One of \code{summary}, ...}

\item{sweep}{If TRUE, call \code{sweep} to update jobs database first}
}
\description{
This is a stub for now. See notes in Obsidian, and compare with Anthill INFO/info()
}
