% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_slurmcollie.R
\name{init_slurmcollie}
\alias{init_slurmcollie}
\title{Initialize user parameters for package slurmcollie}
\usage{
init_slurmcollie()
}
\description{
Reads user-set parameters from \verb{~/slurmcollie.yml} and sets them in the
slurmcollie environment \code{slu}.
}
\details{
\code{slurmcollie.yml} is normally created by `set_up_slurmcollie()'. It should contain the following two lines:

\verb{slurmcollie_dir = <path>} \cr
\verb{login_node = 'login1}

This function reads the user parameters and adds them to the environment \code{slu}. It is
automatically run upon loading the package, and may be rerun by the user if parameter
files are changed.

See \code{README} for details on how to set up ssh access to the login node, which is required
for launching batch jobs in Slurm from R.
}
