% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\title{Kill launched Slurm jobs}
\usage{
kill(filter = NULL, force = FALSE, quiet = FALSE)
}
\arguments{
\item{filter}{Specify jobs with one of:
\itemize{
\item a vector of \code{jobids}
\item 'all' for all jobs
\item a named list to filter jobs with. List items are \verb{<field in jdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{force}{If TRUE, kill the jobs even if Slurm scancel fails. This will potentially
leave hanging batch jobs on Slurm that slurmcollie doesn't know about.}

\item{quiet}{If TRUE, don't complain about jobs not found nor report on
killed jobs}
}
\description{
Uses Slurm \code{scancel} to kill jobs. This can't be done via \code{batchtools}, as
there may be registry conflicts between running jobs and attempts to load the
registry for \code{killJobs}, so we're going directly to Slurm.
}
