% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch batch jobs via Slurm on Unity}
\usage{
launch(
  call,
  args,
  reps,
  resources,
  regdir = the$regdir,
  jobids,
  replace = TRUE
)
}
\arguments{
\item{call}{Name of function to call}

\item{args}{Named list of arguments to called function}

\item{reps}{Vector, list or data frame to vectorize call over}

\item{resources}{Named list of resources, overriding defaults in
\code{batchtools.conf}}

\item{regdir}{Directory containing \code{batchtools} registries}

\item{jobids}{ids for these jobs in jobs database}

\item{replace}{If TRUE, replace existing job ids in jobs database;
otherwise throw an error for existing jobs}
}
\description{
Updates jobs database, \code{the$jdb} to track jobs.
}
