% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_test_finish.R
\name{launch_test_finish}
\alias{launch_test_finish}
\title{a test finish function for launch and sweep}
\usage{
launch_test_finish(jobid, status)
}
\arguments{
\item{jobid}{Job ids to finish for}

\item{status}{Job status}
}
\description{
a test finish function for launch and sweep
}
