% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_database.R
\name{load_database}
\alias{load_database}
\title{Load the specified database unless we've already got it}
\usage{
load_database(database, force = FALSE)
}
\arguments{
\item{database}{Name of database (should \code{db})}

\item{force}{If TRUE, always loads the database, replacing the current one if it exists}
}
\description{
Loads the database from directory \code{the$dbdir} into environment \code{the} if it hasn't
already been loaded or if \code{force = TRUE}.
}
