% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_slu_database.R
\name{load_slu_database}
\alias{load_slu_database}
\title{Load the specified slurmcollie database unless we've already got it}
\usage{
load_slu_database(database = "jdb")
}
\arguments{
\item{database}{Name of database (should be \code{jdb})}
}
\description{
Loads the database from directory \code{slu$dbdir} into environment \code{slu} if it hasn't
already been loaded or if \code{force = TRUE}.
}
