% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile}
\alias{logfile}
\title{Return the names of the log file for a batch job}
\usage{
logfile(jobid)
}
\arguments{
\item{jobid}{A single \code{slurmcollie} batch jobid}
}
\value{
A named list with:
\item{now}{The path to the current log file for the job, which will be in the \code{batchtools}
registry while the job is running, and the \code{slurmcollie} logs folder once it's been marked as
done by \code{slurmcollie}}
\item{done}{The path to the job's log file once the job has been registered as \code{done} by
\code{slurmcollie}}
}
\description{
Logs for \code{slurmcollie} jobs are buried in a \code{batchtools} registry while the job is running,
and moved to the \code{slurmcollie} logs directory once the job is registered as done. This function
points to the current location of the log as well as the final location (where it may not yet
be). You can use this for easy monitoring of logs via \code{cat} in Linux or by opening in an \code{SFTP}
client. Use \link{showlog} to simply print the current log.
}
