% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Build a geoTIFF map of predictions for specified model fit}
\usage{
map(fit, clip = NULL, result = NULL)
}
\arguments{
\item{clip}{Optional clip, vector of \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax}}

\item{result}{Optional result filename. If not provided, uses name from database if it exists. Otherwise, constructs a name.}

\item{model}{Model ID, fit filename, or fit object - figure it out)}
}
\description{
Console command to launch a prediction run via \code{do_map}, typically in a batch job on Unity.
}
