% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{msg}
\alias{msg}
\title{Append message to the log file and also write it to the display}
\usage{
msg(message, logfile)
}
\arguments{
\item{message}{text of message}

\item{logfile}{path and name of log file}
}
\description{
Displays message to the console, writing it to the log with a timestamp
if a log file is provided.
}
\keyword{internal}
