% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_slurmcollie_db.R
\name{new_slurmcollie_db}
\alias{new_slurmcollie_db}
\title{Create a new jobs database for the slurmcollie package}
\usage{
new_slurmcollie_db(database = "jdb", really = FALSE)
}
\arguments{
\item{database}{Name of database (should be \code{jdb})}

\item{really}{If TRUE, creates database, \strong{destroying existing database}}
}
\description{
Creates a new empty jobs database (\code{jdb}). This is a
drastic function, intended to be used only when initially creating a database
or when an existing database is a hopeless mess. Use with great care--this
function will destroy any existing database and backups. When creating a new
jobs database (\code{jdb}), it also deletes all logs and registries, and saves an
empty database to disk immediately. \strong{This function is drastic and
unrecoverable.}
}
