% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars}
\alias{pars}
\title{Return path to parameter file}
\usage{
pars(file)
}
\arguments{
\item{file}{Parameter file name}
}
\value{
Path and name of parameter file
}
\description{
Given a file name, return path to it in user parameters directory
}
\keyword{internal}
