% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickvars.R
\name{pickvars}
\alias{pickvars}
\title{pick variables based on importance >= n}
\usage{
pickvars(n, importance = import)
}
\arguments{
\item{n}{Cutoff value for variable importance.}

\item{importance}{Variable importance object from \code{fit}}
}
\description{
pick variables based on importance >= n
}
