% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge}
\alias{purge}
\title{Purge selected jobs from the jobs database}
\usage{
purge(filter = NULL, force = FALSE, quiet = FALSE)
}
\arguments{
\item{filter}{Specify jobs with one of:
\itemize{
\item a vector of \code{jobids}
\item 'all' for all jobs
\item a named list to filter jobs with. List items are \verb{<field in jdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{force}{If TRUE, purges jobs even if they're pending, queued, or running - don't
do this casually!}

\item{quiet}{If TRUE, don't chatter}
}
\description{
Purges jobs and their logs.
}
