% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pars_table.R
\name{read_pars_table}
\alias{read_pars_table}
\title{Read a parameter table}
\usage{
read_pars_table(name)
}
\arguments{
\item{name}{Parameter name}
}
\value{
Data frame of the parameter file
}
\description{
Given a parameter name the corresponds to tab-delimited file, read and return the file
}
\keyword{internal}
