% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample Y and X variables for a site}
\usage{
sample(
  site,
  pattern = "",
  n = NULL,
  p = NULL,
  distance = NULL,
  classes = NULL,
  balance = TRUE,
  balance_excl = c(7, 33),
  result = NULL,
  transects = NULL,
  xy = FALSE
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Default = all sites.}

\item{pattern}{Regex filtering rasters of predictor variables, case-insensitive.
Default = "" (match all). Note: only files ending in \code{.tif} are included in any case.}

\item{n}{Number of total samples to return.}

\item{p}{Proportion of total samples to return. Use p = 1 to sample all.}

\item{distance}{Sample points with an average spacing of \code{distance} m. Sampling is
done in cells, so there's no guarantee points will be separated by this much.}

\item{classes}{Class or vector of classes in transects to sample. Default is all
classes.}

\item{balance}{If TRUE, balance number of samples for each class. Points will be randomly
selected to match the sparsest class.}

\item{balance_excl}{Vector of classes to exclude when determining sample size when
balancing. Include classes with low samples we don't care much about.}

\item{result}{Name of result file. If not specified, file will be constructed from
site, number of X vars, strategy, and distance.}

\item{transects}{Name of transects file; default is \code{transects}.}

\item{xy}{If TRUE, also get the X and Y coordinates of each sample point.}
}
\value{
Sampled data table (invisibly)
}
\description{
There are three mutually available sampling strategies (n, p, and distance). You
must choose exactly one.
}
