% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_database.R
\name{save_database}
\alias{save_database}
\title{Save the specified database}
\usage{
save_database(database)
}
\arguments{
\item{database}{Name of database (\code{jdb} or \code{db})}
}
\description{
Saves the database in directory \code{the$dbdir}. Previous versions are renamed e.g., \code{jbd_1.RDS},
\code{jbd_2.RDS}, etc. Will need a mechanism to delete databases more than a week old (or something).
}
