% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dirs.R
\name{set_dirs}
\alias{set_dirs}
\title{Set complete paths to all directories}
\usage{
set_dirs()
}
\description{
Sets all directory names from \code{pars.yml} or defaults. Note that \code{basedir}, \code{parsdir},
and \code{scratchdir} are set by \code{\link[=init]{init()}}.
}
