% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_slurmcollie.R
\name{set_up_slurmcollie}
\alias{set_up_slurmcollie}
\title{Set up files for package slurmcollie}
\usage{
set_up_slurmcollie(directory = NULL, login_node = "login1", replace = FALSE)
}
\arguments{
\item{directory}{The user directory for slurmcollie files}

\item{login_node}{The name of a login node with ssh access}

\item{replace}{If TRUE, replaces \code{slurmcollie.yml} and the template files if any exist}
}
\description{
Creates directories and a user parameter file for the \code{slurmcollie} package.
\code{slurmcollie.yml} is normally created by `set_up_slurmcollie()'. It should contain the following two lines:
}
\details{
\verb{slurmcollie_dir = <path>} \cr
\verb{login_node = 'login1}

This function reads the user parameters and adds them to the environment \code{slu}. It is
automatically run upon loading the package, and may be rerun by the user if parameter
files are changed.

This function also creates the directory named in \code{directory} and subdirectories
\code{databases}, \code{template}, \verb{registries', and }logs\verb{if they don't already exist. It's okay to share these directories with other R code; for instance, the}saltmarsh\verb{package also uses}databases`.

See \code{README} for details on how to set up ssh access to the login node, which is required
for launching batch jobs in Slurm from R.
}
