% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showlog.R
\name{showlog}
\alias{showlog}
\title{Prints the contents log file for a batch job}
\usage{
showlog(jobid)
}
\arguments{
\item{jobid}{A single \code{slurmcollie} batch jobid}
}
\description{
Logs for \code{slurmcollie} jobs are buried in a \code{batchtools} registry while the job is running,
and moved to the \code{slurmcollie} logs directory once the job is registered as done. This function
prints the current log file the job. Use \link{logfile} to get the current and final path for the
job's log file.
}
