% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab.R
\docType{data}
\name{slu}
\alias{slu}
\title{Script to set up environment for parameters in package slurmcollie}
\format{
An object of class \code{environment} of length 6.
}
\usage{
slu
}
\description{
The environment \code{slu} is used for user parameters. They
are assigned by \code{init_slurmcollie()}, which is run when the package is
loaded, and also may be run by the user, e.g., which parameter
files are changed.
}
\keyword{datasets}
