% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{sweep}
\alias{sweep}
\title{Sweep up finished batch jobs}
\usage{
sweep(registriesdir = the$registriesdir)
}
\arguments{
\item{registriesdir}{Directory containing \code{batchtools} registries}
}
\description{
Checks status of all jobs in all registries. Fills database with run stats, then deletes all
registries that have been completely swept. Reports how many jobs have been swept, how many
successful, how many failed, and how many are still outstanding. If all registries have been
swept, we'll get to start over at reg001. Saves the database, of course!
}
