% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep_jobs.R
\name{sweep_jobs}
\alias{sweep_jobs}
\title{Sweep up finished batch jobs}
\usage{
sweep_jobs(stats = TRUE, quiet = FALSE)
}
\arguments{
\item{stats}{If TRUE, fills in run stats}

\item{quiet}{If TRUE, don't say anything; otherwise does info('summary') at the end}
}
\description{
Checks status of all jobs in all registries. Fills database with run stats, then deletes all
registries that have been completely swept. Reports how many jobs have been swept, how many
successful, how many failed, and how many are still outstanding. If all registries have been
swept, we'll get to start over at reg001. Saves the database, of course!
}
\details{
Calls \code{finish} functions passed to \code{launch}, to, for example, update a parent database. These
are called for any newly-done jobs, whether they were successful or not. See \link{launch} for
details. If finish functions don't exist in the workspace (i.e., their package isn't loaded),
you'll get a warning. You'll need to load the finish functions and rerun \code{info}. Any actions
the finish functions are supposed to do (e.g., update a database) will be left undone until
you attend to this.

Getting run stats (memory, CPU time, etc.) take about 75\% of the execution time, so if you
don't need them and are impatient, use \code{stats = FALSE}.

Since \code{info} calls \code{sweep_jobs} by default, \code{info} is the normal user-facing function for updating
the jobs database, though calling \code{sweep_jobs} is functionally equivalent to calling \code{info()}, so
take your pick.
}
