% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{the}
\alias{the}
\title{Script to set up environment for parameters}
\format{
An object of class \code{environment} of length 20.
}
\usage{
the
}
\description{
The environment \code{the} is used for user parameters. They
are assigned by \code{init()}, which is run when the package is
loaded, and also may be run by the user, e.g., which parameter
files are changed.
}
\keyword{datasets}
