% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pars_table.R
\name{write_pars_table}
\alias{write_pars_table}
\title{Write a parameter table}
\usage{
write_pars_table(data, name)
}
\arguments{
\item{data}{Data frame to write}

\item{name}{Parameter name}
}
\description{
Given a parameter name the corresponds to tab-delimited file, write a data frame to it.
}
\keyword{internal}
