% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files.R
\name{check_files}
\alias{check_files}
\title{Check that each source file exists on the result directory and is up to date}
\usage{
check_files(files, gd, sourcedir, resultdir)
}
\arguments{
\item{files}{Vector of files to check}

\item{gd}{Source Drive info (optional), named list of
\itemize{
\item \code{dir}            Google directory info, from \link{get_dir}
\item \code{sourcedrive}    which source drive (\code{local}, \code{google}, or \code{sftp})
\item \code{sftp}           \code{list(url, user)}
}}

\item{sourcedir}{Origin directory of files}

\item{resultdir}{Target directory of files - see if origin files are here and up to date}
}
\value{
A vector corresponding to files, TRUE for those that are up to date
}
\description{
Check that each source file exists on the result directory and is up to date
}
