% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_jobs.R
\name{filter_jobs}
\alias{filter_jobs}
\title{Filter jobs database}
\usage{
filter_jobs(filter)
}
\arguments{
\item{filter}{Specify jobs with one of:
\itemize{
\item a vector of \code{jobids}
\item 'all' for all jobs
\item a named list to filter jobs. List items are \verb{<field> = <value>}, where
\verb{<field>} is a field in the jobs database, and \verb{<value>} is a regex for
character fields, or an actual value (or vector of values) for logical
or numeric fields.
}}
}
\value{
A vector of rows numbers in \code{slu$jdb}
}
\description{
Filter jobs database
}
