% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_job_id.R
\name{get_job_id}
\alias{get_job_id}
\title{Get Slurm job id for a \code{batchtools} job}
\usage{
get_job_id(job.id, reg = getDefaultRegistry())
}
\arguments{
\item{job.id}{The job.id of a \code{batchtools} job}

\item{reg}{\code{batchtools} registry object}
}
\description{
Gets the Slurm job id for the specified \code{batchtools} job id in the specified registry.
Throws an error if the \code{batchtools} job or registry don't exist. If there is no Slurm
job id yet, returns NA.
}
\details{
In order for this to work, you must add the following line to your Slurm template
(\code{slurm.tmpl}): \cr \cr
\verb{echo $SLURM_JOB_ID > <\%= log.file \%>.slurmjobid}
}
