% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Give info on batch jobs}
\usage{
info(
  columns = "normal",
  filter = "all",
  sort = "jobid",
  decreasing = FALSE,
  nrows = NA,
  summary = TRUE,
  table = TRUE,
  sweep = TRUE,
  timezone = "America/New_York"
)
}
\arguments{
\item{columns}{Specifies which columns to include in the jobs table. May be
one of
\itemize{
\item \emph{brief} (1) includes \code{jobid}, \code{status}, \code{error}, \code{comment}
\item \emph{normal} (2)  includes \code{jobid}, \code{launched}, \code{call}, \code{rep}, \code{local}, \code{status},
\code{error}, \code{cores}, \code{cpu}, \code{cpu_pct}, \code{mem_req}, \code{mem_gb}, \code{walltime}, \code{comment}
\item \emph{long} (3) includes \code{jobid}, \code{launched}, \code{call}, \code{rep}, \code{local}, \code{sjobid}, \code{status},
\code{state}, \code{reason}, \code{error}, \code{message}, \code{done}, \code{cores}, \code{mem_req}, \code{mem_gb}, \code{walltime},
\code{cpu}, \code{cpu_pct}, \code{log}, \code{comment}
\item \emph{all} (4) includes all columns
\item 1, 2, 3, or 4 is a shortcut for the above column sets
\item A vector of column names to include
}}

\item{filter}{Specify jobs with one of:
\itemize{
\item a vector of \code{jobids}
\item 'all' for all jobs
\item a named list to filter jobs with. List items are \verb{<field in jdb> = <value>},
where \verb{<value>} is a regex for character fields, or an actual value (or vector of
values) for logical or numeric fields.
}}

\item{sort}{The name of the column to be used to sort the table}

\item{decreasing}{If TRUE, sort in descending order}

\item{nrows}{Number of rows to display in the jobs table. Positive numbers
display the first \emph{n} rows, and negative numbers display the last \emph{n} rows.
Use \code{nrows = NA} to display all rows.}

\item{summary}{If TRUE, displays jobs summary}

\item{table}{If TRUE, displays jobs table}

\item{sweep}{If TRUE, call \code{sweep_jobs} to update jobs database first}

\item{timezone}{Time zone for launch time; use NULL to leave times in native UTC}
}
\value{
The processed jobs table, invisibly
}
\description{
Provides information on jobs in the batch jobs database. Options allow for
selecting columns, filtering, sorting, and displaying a brief summary or a
complete table. Normally, \code{sweep_jobs} is called up front to update the database,
so there may be a delay of a second or two.
}
\details{
Note that memory is always reported in GB.

Also note that local calls will not record \code{cores}, \code{cpu}, \code{cpu_pct}, nor \code{mem_req}.
They do record mem_gb and walltime.
}
