% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize salt-marsh-mapping with user parameters}
\usage{
init()
}
\description{
Reads user-set parameters for package saltmarsh.
}
\details{
User parameters are set in two distinct locations:
\enumerate{
\item The initialization file, in the user's home directory, \verb{~/saltmarsh.yml}.
This file should contain two lines:

\verb{parsdir: c:/Work/etc/saltmarsh/pars}

\code{parsfile: pars.yml}

a. `parsdir`` points to the parameter directory

b. \code{parsfile} points to the main parameter file
\item Everything else, in \code{parsdir}. The primary parameter file is \code{pars.yml}, which
points to other paramters (such as sites.txt).
}

This approach splits the user- and platform-dependent parameters (saltmarsh.yml)
from parameters that are likely to be shared among users and platforms (those in
\code{parsdir}). It allows multiple users on a shared machine (such as Unity cluster)
to set user-specific parameters if need be, while sharing other parameters.

This function reads the user parameters and sets a local environment \code{the} with
all parameters. It is automatically run upon loading the package, and may be
rerun by the user if parameter files are changed.
}
