% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\title{Kill launched Slurm jobs}
\usage{
kill(jobids, quiet = FALSE)
}
\arguments{
\item{jobids}{One or more job ids}

\item{quiet}{If TRUE, don't complain about jobs not found nor report on killed jobs}
}
\description{
Uses Slurm \code{scancel} to kill jobs. This won't work for jobs that haven't reported a Slurm JobID yet,
so you'll need to run \code{sweep} before running this. It also fail for jobs that haven't reported back
to batchtools.
}
\details{
This can't be done via \code{batchtools}, as there may be registry conflicts between running jobs and
attempts to load the registry for \code{killJobs}, so we're going directly to Slurm.
}
