% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.R
\name{kill}
\alias{kill}
\title{Kill launched Slurm jobs}
\usage{
kill(jobids = NULL, filter = NULL, quiet = FALSE)
}
\arguments{
\item{jobids}{One or more job ids}

\item{filter}{A named list to filter jobs by columns in the jobs database.
List items are \verb{<field in jdb> = <value>}, where \if{html}{\out{<value>}} is a regex for
character fields, or an actual value (or vector of values) for logical or
numeric fields.}

\item{quiet}{If TRUE, don't complain about jobs not found nor report on
killed jobs}
}
\description{
Uses Slurm \code{scancel} to kill jobs. This can't be done via \code{batchtools}, as
there may be registry conflicts between running jobs and attempts to load the
registry for \code{killJobs}, so we're going directly to Slurm.
}
