% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Launch batch jobs via Slurm on Unity}
\usage{
launch(
  call,
  args,
  reps = 1,
  argname = "rep",
  moreargs = list(),
  resources = list(),
  regdir = the$regdir,
  jobids = NULL,
  comment = "",
  finish = NA,
  replace = TRUE
)
}
\arguments{
\item{call}{Name of function to call}

\item{args}{Named list of arguments to called function}

\item{reps}{Vector, list, or data frame to vectorize call over. If a
named list or data frame, the names must correspond to the fuction's
arguments. If a vector or unnamed list, \code{argname} is used.}

\item{argname}{Name of \code{reps} argument in function to be called, used
only when \code{reps} is a vector or unnamed list}

\item{moreargs}{a named list of additional arguments to the called function,
not vectorized over}

\item{resources}{Named list of resources, overriding defaults in
\code{batchtools.conf}}

\item{regdir}{Directory containing \code{batchtools} registries}

\item{jobids}{ids for these jobs in jobs database. Supply existing
jobs to relaunch jobs; NA or NULL will create new jobs.}

\item{comment}{Optional comment; will be recycled for multiple reps}

\item{finish}{Optional name of a function to run for completed jobs,
for example \code{finish = 'sweep_fit'} to gather fit stats}

\item{replace}{If TRUE, replace existing job ids in jobs database;
otherwise throw an error for existing jobs}
}
\description{
Updates jobs database, \code{the$jdb} to track jobs.
}
