% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_db.R
\name{new_db}
\alias{new_db}
\title{Create a new database}
\usage{
new_db(database, really = FALSE)
}
\arguments{
\item{database}{Name of database (\code{jdb} or \code{db})}

\item{really}{If TRUE, creates database, \strong{destroying existing database}}
}
\description{
Creates a new empty fits database (\code{db}) or jobs database (\code{jdb}). Use with great care--this
function will destroy any existing database. It saves the results to disk immediately.
}
