% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_dir.R
\name{resolve_dir}
\alias{resolve_dir}
\title{Resolve directory with embedded \verb{<site>}}
\usage{
resolve_dir(dir, site)
}
\arguments{
\item{dir}{Directory path}

\item{site}{Site code (3 letter abbreviation)}
}
\value{
Directory path including specified site.
}
\description{
Resolve directory with embedded \verb{<site>}
}
\keyword{internal}
