% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample Y and X variables for a site}
\usage{
sample(
  site,
  pattern = "",
  n = NULL,
  p = NULL,
  d = NULL,
  classes = NULL,
  balance = TRUE,
  balance_excl = c(7, 33),
  result = NULL,
  transects = NULL,
  drop_corr = NULL,
  reuse = FALSE
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Default = all sites.}

\item{pattern}{Regex filtering rasters of  r predictor variables, case-insensitive.
Default = "" (match all). Note: only files ending in \code{.tif} are included in any case.}

\item{n}{Number of total samples to return.}

\item{p}{Proportion of total samples to return. Use p = 1 to sample all.}

\item{d}{Mean distance in cells between samples. No minimum spacing is guaranteed.}

\item{classes}{Class or vector of classes in transects to sample. Default is all
classes.}

\item{balance}{If TRUE, balance number of samples for each class. Points will be randomly
selected to match the sparsest class.}

\item{balance_excl}{Vector of classes to exclude when determining sample size when
balancing. Include classes with low samples we don't care much about.}

\item{result}{Name of result file. If not specified, file will be constructed from
site, number of X vars, and strategy.}

\item{transects}{Name of transects file; default is \code{transects}.}

\item{drop_corr}{Drop one of any pair of variables with correlation more than \code{drop_corr}.}

\item{reuse}{Reuse the named file (ending in \verb{_all.txt}) from previous run, rather
than resampling. Saves a lot of time if you're changing \code{n}, \code{p}, \code{d}, \code{balance},
\code{balance_excl}, or \code{drop_corr}.}
}
\value{
Sampled data table (invisibly)
}
\description{
There are three mutually available sampling strategies (n, p, and d). You
must choose exactly one. \code{n} samples the total number of points provided.
\code{p} samples the proportion of total points (after balancing, if \code{balance} is
selected. \code{d} samples points with a mean (but not guaranteed) minimum distance.
}
\details{
Results are saved in four files:
\enumerate{
\item \if{html}{\out{<result>}}_all.txt - A text version of the full dataset (selected by \code{pattern}
but not subsetted by \code{n}, \code{p}, \code{d}, \code{balance}, or \code{drop_corr}). Readable by
any software.
\item \if{html}{\out{<result>}}_all.RDS - An RDS version of the full dataset; far faster to read
in R (1.1 s vs. 14.4 s in one example).
\item \if{html}{\out{<result>}}.txt - A text version of the final selected and subsetted dataset,
as a text file.
\item \if{html}{\out{<result>}}.RDS - An RDS version of the final dataset.
}

\strong{Memory requirements: I've measured up to 28.5 GB.}
}
