% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Sample Y and X variables for a site}
\usage{
sample(
  site,
  pattern = "",
  strategy = "all",
  distance = 1,
  classes = NULL,
  balance = TRUE,
  result = NULL,
  transects = NULL,
  xy = FALSE
)
}
\arguments{
\item{site}{One or more site names, using 3 letter abbreviation. Default = all sites.}

\item{pattern}{Regex filtering rasters of predictor variables, case-insensitive.
Default = "" (match all). Note: only files ending in \code{.tif} are included in any case.}

\item{strategy}{Sampling strategy. One of
\enumerate{
\item all - sample all points with field data. Use this when assessing spatial
autocorrelation, or if don't believe it's a thing.
\item distance - sample points with an average spacing of \code{distance} m. Sampling is
done by cells, so there's no guarantee points will be separated by this much.
}}

\item{distance}{Average spacing (m) between points.}

\item{classes}{Class or vector of classes in transects to sample. Default is all
classes.}

\item{balance}{If TRUE, balance number of samples for each class. Points will be randomly
selected to match the sparsest class.}

\item{result}{Name of result file. If not specified, file will be constructed from
site, number of X vars, strategy, and distance.}

\item{transects}{Name of transects file; default is \code{transects}.}

\item{xy}{If TRUE, also get the X and Y coordinates of each sample point.}
}
\description{
Sample Y and X variables for a site
}
