% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_slu_database.R
\name{save_slu_database}
\alias{save_slu_database}
\title{Save the specified slurmcollie database}
\usage{
save_slu_database(database)
}
\arguments{
\item{database}{Name of database (should be \code{jdb})}
}
\description{
Saves the database in directory \code{slu$dbdir}. Previous versions are renamed e.g., \code{jdb_1.RDS},
\code{jdb_2.RDS}, etc and moved to \verb{backups/}. Everything in \verb{backups/} is save to purge if you're
hurting for disk space.
}
